Move findHeadInCache and add unit test
